+++ ftp/src/client.d/cl_server_write.c
@@ -6,7 +6,7 @@
 /*   By: gbourgeo <gbourgeo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/01 21:08:30 by gbourgeo          #+#    #+#             */
-/*   Updated: 2020/02/02 02:14:52 by gbourgeo         ###   ########.fr       */
+/*   Updated: 2020/02/03 17:31:05 by gbourgeo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int			cl_server_write(const char *buf, int len, t_server *sv)
 		*sv->wr.tail++ = buf[i++];
 		if (sv->wr.tail >= sv->wr.buff + sizeof(sv->wr.buff))
 			sv->wr.tail = sv->wr.buff;
-		if (sv->wr.len++ >= sizeof(sv->wr.buff))
+		if (sv->wr.len++ >= (int)sizeof(sv->wr.buff))
 			if ((errnb = cl_server_send(sv->fd_ctrl, &sv->wr)) != IS_OK)
 				return (errnb);
 	}
